Factor out handle auth success
